%
%  Copyright (C) 2011-2015 Christina Yassouridis
% 
%


\name{bones}
\docType{data}
\alias{bones}
\title{Bone density}
\description{
  Relative spinal bone mineral density measurements on 261 North
  American adolescents.  Each value is the difference in mineral density
  taken on two consecutive visits, divided by the average. The age is
  the average age over the two visits.
}
\usage{data("bones")}
\format{General format \code{"Format1"} (see \code{\link{formatFuncy}}): A matrix of dimension 261x3.}
\source{http://statweb.stanford.edu/~tibs/ElemStatLearn/}
\references{
  Hastie, Tibshirani and Friedman. The Elements of Statistical Learning (2nd edition)
  (2009). Springer-Verlag. 763 pages. 
  
  Gareth M. James and Trevor J. Hastie. Functional Linear Discriminant
  Analysis for Irregularly Sampled Curves. Journal of the Royal
  Statistical Society Series B. 63. 533--550. 2001

  Gareth M. James and Catherine A. Sugar. Clustering for Sparsely Sampled Functional Data
  . Journal of the American Statistical Association. 98. 397--408. 2003
}
\keyword{datasets}
