%
%  Copyright (C) 2011-2015 Christina Yassouridis
% 
%

\name{CClust-methods}
\Rdversion{1.1}
\docType{methods}
\title{Methods for CClust object. }
\alias{[,CClustO,ANY,ANY,ANY-method}
\alias{[,CClustRes,ANY,ANY,ANY-method}
\alias{[<-,CClustO,ANY,ANY,ANY-method}
\alias{[<-,CClustRes,ANY,ANY,ANY-method}
\alias{summary,CClustRes-method}

\description{Mehods for CClust objects.}

\arguments{
  Not documented, for internal use only.
}

\value{
  Not documented, for internal use only.
}

\author{Christina Yassouridis}
