%
%  Copyright (C) 2011-2015 Christina Yassouridis
% 
%

\name{electricity}
\docType{data}
\alias{electricity}
\title{Electric power}
\description{The electric power home consumers at Buenos Aires,
Argentina in 2001. For 104 households, measurements were taken at each of the 96 subintervals
of 15 min in every week day, Monday to Friday, during January 2001.
}
\usage{data("electricity")}
\format{Regular format \code{"Format2"} (see \code{\link{formatFuncy}}): A matrix of dimension 261x104.}
\references{
Juan Antonio Cuesta-Albertos and Ricard Fraiman. Impartial trimmed
k-means for functional data. Computational Statistics & Data
Analysis. 51. 4864--4877. 2007
}
\keyword{datasets}
