%
%  Copyright (C) 2011-2015 Christina Yassouridis
% 
%


\name{genes}
\docType{data}
\alias{genes}
\title{Gene expression profiles}
\description{
  A subset of the Drosophila life cycle gene expression data of Arbeitman
  et al. (2002). The original data set contains 77 gene expression
  profiles during 58 sequential time points from the embryonic, larval,
  and pupal periods of the life cycle.
}
\usage{data("genes")}
\format{Data in format "Format2" (see \code{\link{formatFuncy}}). A matrix of dimension 44x77, a cluster vector of length 44.}
\source{GDS191 in Gene Expression Ominibus http://www.ncbi.nlm.nih.gov/geo/}
\references{
  J.-M. Chiou and P.-L. Li J. R.
  Functional clustering and identifying substructures of longitudinal
  data. Statist. Soc. B. Volume 69. 679–699. 2007

  Arbeitman, M.N and al. Gene expression during the life cycle of Drosophila
  melanogaster. Science. 297(5590). 2002
}
\keyword{datasets}
