%
%  Copyright (C) 2011-2015 Christina Yassouridis
% 
%

\name{lowflow}
\docType{data}
\alias{lowflow}
\title{Gauges in Austria}

\description{The dataset consists of 82 gauges in Upper Austria where
  streamflow minima were identified each winter during the years (1976-2008). \code{lowflow} is a
  list of data in format "Format2" (see \code{\link{formatFuncy}}) and a
  location matrix of measurements. 
}
\usage{data("lowflow")}
\format{Regular format \code{"Format2"} (see \code{\link{formatFuncy}}):
  A matrix of dimension 33x82.}
\references{
  G. Laaha, D. Koffler, K. Haslinger, W. Schöner, J. Parajka, A. Viglione, J. Zehetgruber, C. Yassouridis, and G. Blöschl. Assessing climate impacts on low flows and droughts combining upward and downward approaches.
  10th International Conference on Geostatistics for Environmental
  Applications. JUL 9-11. Paris, 2014
}
\keyword{datasets}
